# docs/swagger.yaml
basePath: /
definitions:
  models.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.OTPRequest:
    properties:
      phone_number:
        type: string
    type: object
  models.OTPVerifyReq:
    properties:
      code:
        type: string
      phone_number:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      registration_data:
        type: string
      updated_at:
        type: string
    type: object
  models.UserListResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple OTP-based authentication service
  title: OTP Service API
  version: "1.1"
paths:
  /api/v1/auth/send-otp:
    post:
      consumes:
      - application/json
      description: Send a 6 digit OTP code to the specific phone number
      parameters:
      - description: Phone number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.OTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send otp to phone number
      tags:
      - auth
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get a paginated users list with optional search
      parameters:
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - description: Search by phone number
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get list of users
      tags:
      - users
  /api/v1/users/{id}:
    get:
      consumes:
      - application/json
      description: get a single user by speicified ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - users
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verify the OTP code and return JWT token
      parameters:
      - description: Phone number and OTP code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.OTPVerifyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify OTP and authenticate user
      tags:
      - auth
swagger: "2.0"